+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `ICanBeReadOnly` interface for scriptable objects that can be marked as read-only
 - `INotifyScriptableCollectionChanged` interface for scriptable collections that can be used to track collection changes
 - Full parity with `List<T>` to `ScriptableList<T>`
+- `ScriptableValueListener<T>` now has `OnBeforeValueChangingInvoked(T oldValue, T newValue)`, `OnAfterValueChangingInvoked(T oldValue, T newValue)`, `OnBeforeValueChangedInvoked(T oldValue, T newValue)`, and `OnAfterValueChangedInvoked(T oldValue, T newValue)` overridable methods
 
 ### Changed
 
@@ -33,7 +34,8 @@
 - `ScriptableDictionary<TKey, TValue>.TrimExcess()` can now be called even when the object is marked as read-only
 - `ScriptableDictionary<TKey, TValue>` now implements `INotifyCollectionChanged` and `InotifyScriptableCollectionChanged<KeyValurPair<TKey, TValue>>`
 - **BREAKING**: `ScriptablePool<T>.Return(T item)` is now obsolete, use `Release(T item)` instead
-- **BREAKING**: `ScriptableValueListener.ToggleListening(bool listen)` is now obsolete, use `SetListening(bool listen)` instead
+- **BREAKING**: `ScriptableListenerBase.ToggleListening(bool listen)` is now obsolete, use `SetListening(bool listen)` instead
+- **BREAKING**: `ScriptableValueListener<T>`'s `OnCurrentValueChanging`, `OnCurrentValueChanged`, `SetTargetValue`, and `SetListening` methods can no longer be overriden
 
 ### Fixed